code corrections.

+++ servus/test/servus.h
@@ -103,7 +103,7 @@ namespace servus
                     ValueMap& values = _instanceMap[i->_instance];
                     values.clear();
                     values["servus_host"] = "localhost";
-                    values["servus_port"] = std::to_string(_port);
+                    values["servus_port"] = std::to_string(static_cast<unsigned int>(_port));
 
                     for (const auto& j : i->_data)
                         values[j.first] = j.second;
